Another round of refactoring for cleaning the server code
As title of the commit
