+++ server/package.json
@@ -17,12 +17,15 @@
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
+    "eslint": "^8.52.0",
     "express": "^4.18.2",
     "express-session": "^1.18.1",
     "express-socket.io-session": "^1.3.5",
     "groq-sdk": "^0.15.0",
     "mongoose": "^8.10.1",
     "node-forge": "^1.3.1",
+    "prettier": "^3.1.0",
+    "prettier-eslint": "^16.1.2",
     "socket.io": "^4.7.1",
     "socket.io-express-session": "^0.1.3"
   },
